Join HoeffdingAdaptiveTreeClassifier nodes
